Conflicts:
	Applications/EnhanceUsingDiscriminantAnalysis/EnhanceUsingDiscriminantAnalysis.cxx
	Applications/EnhanceUsingDiscriminantAnalysis/EnhanceUsingDiscriminantAnalysis.xml
	Applications/EnhanceUsingNJetDiscriminantAnalysis/EnhanceUsingNJetDiscriminantAnalysis.cxx
	Applications/EnhanceUsingNJetDiscriminantAnalysis/EnhanceUsingNJetDiscriminantAnalysis.xml
	Applications/SegmentTubeSeeds/CMakeLists.txt
	Applications/SegmentTubeSeeds/SegmentTubeSeeds.xml
	Base/CLI/Testing/tubeCLIHelperFunctionsTest.cxx
	Base/CLI/tubeCLIHelperFunctions.h
	Base/Common/tubeTestMain.h
	Base/Filtering/Testing/tubeBaseFilteringHeaderTest.cxx
	Base/Filtering/Testing/tubeBaseFilteringPrintTest.cxx
	Base/Filtering/Testing/tubeBaseFilteringTests.cxx
	Base/IO/Testing/tubeBaseIOPrintTest.cxx
	Base/IO/Testing/tubeBaseIOTests.cxx
	Base/IO/itkTubeMetaLDA.cxx
	Base/IO/itkTubeMetaLDA.h
	Base/IO/itkTubeMetaNJetLDA.cxx
	Base/IO/itkTubeMetaNJetLDA.h
	Base/Numerics/CMakeLists.txt
	Base/Numerics/Testing/CMakeLists.txt
	Base/Numerics/Testing/itkTubeNJetFeatureVectorGeneratorTest.cxx
	Base/Numerics/Testing/itkTubeNJetLDAGenerator2FTest.cxx
	Base/Numerics/Testing/itkTubeSupervisedLinearBasisGeneratorTest.cxx
	Base/Numerics/Testing/tubeBaseNumericsHeaderTest.cxx
	Base/Numerics/Testing/tubeBaseNumericsPrintTest.cxx
	Base/Numerics/Testing/tubeBaseNumericsTests.cxx
	Base/Numerics/itkTubeLDAGenerator.h
	Base/Numerics/itkTubeLDAGenerator.txx
	Base/Numerics/itkTubeNJetImageFunction.txx
	Base/Numerics/itkTubeNJetLDAGenerator.h
	Base/Numerics/itkTubeNJetLDAGenerator.txx
	Base/Numerics/itkTubeNJetLDAGenerator2F.cxx
	Base/Numerics/itkTubeNJetLDAGenerator2F.h
	Base/Segmentation/Testing/itkTubeRidgeSeedSupervisedLinearBasisGeneratorTest.cxx
	Base/Segmentation/Testing/tubeBaseSegmentationHeaderTest.cxx
	Base/Segmentation/Testing/tubeBaseSegmentationPrintTest.cxx
	Base/Segmentation/Testing/tubeBaseSegmentationTests.cxx
	Base/Segmentation/itkTubeLabelOverlapMeasuresImageFilter.h
	Base/Segmentation/itkTubeLabelOverlapMeasuresImageFilter.txx
	Base/Segmentation/itkTubeOtsuThresholdMaskedImageFilter.h
	Base/Segmentation/itkTubeOtsuThresholdMaskedImageFilter.txx
	Base/Segmentation/itkTubePDFSegmenter.h
	Base/Segmentation/itkTubeRidgeSeedGenerator.h
	Base/Segmentation/itkTubeRidgeSeedGenerator.txx
	Utilities/Testing/CMakeLists.txt

+++ Base/USTK/Testing/itkLabelMapToAcousticImpedanceImageFilterTest.cxx
@@ -20,17 +20,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 =========================================================================*/
-#include <itkLabelMapToAcousticImpedanceImageFilter.h>
+
+#include "itkLabelMapToAcousticImpedanceImageFilter.h"
 #include <fstream>
 
-#include "itkImageFileReader.h"
-#include "itkImageFileWriter.h"
+#include <itkImageFileReader.h>
+#include <itkImageFileWriter.h>
 
 
 template< class TLookupTable >
 int ReadLookupTableFromCSV( const char * filename, TLookupTable & lookupTable );
 
-int itkLabelMapToAcousticImpedanceImageFilterTest( int argc, char * argv [] )
+int itkLabelMapToAcousticImpedanceImageFilterTest( int argc, char * argv[] )
 {
   // Argument parsing.
   if( argc < 4 )
@@ -57,7 +58,6 @@ int itkLabelMapToAcousticImpedanceImageFilterTest( int argc, char * argv [] )
     AcousticImpedanceImageType;
 
   typedef std::vector< float > LookupTableType;
-
 
 
   // Reader.
@@ -126,7 +126,7 @@ int ReadLookupTableFromCSV( const char * filename, TLookupTable & lookupTable )
     return EXIT_FAILURE;
     }
 
-  size_t label = 0;
+  unsigned int label = 0;
   char tissueType[256];
   float acousticImpedance;
   inputStream >> label;
